media/gpu/fourcc: allow to create Fourcc from an uint32
Currently Fourccs are created from a video pixel format from a specific
set of APIs. However Fourccs are actually universal values that can be
created from an integer, so add a constructor allowing this.

This new constructor only accepts its set of internally-recognized
fourccs, and returns an Optional. If a value is present, it is
guaranteed to be valid.

Since we are using the V4L2 fourccs internally, this allows us to
simplify the constructor from a V4L2 format a bit as well.

Finally, libyuv used to directly compare a Fourcc to an integer - this
is dangerous, so build a second Fourcc using this new constructor to
perform the comparison instead. This will allow us to remove these
dangerous operators in a following CL.

BUG=1031082
BUG=b:143498574
TEST=VDAtests passing on Hana.
TEST=./media_unittests --gtest_filter=FourccTest* passing.

Change-Id: Icfdc8ad037d66c05c0694dca759f7bd8c1be76dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1957193
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723737}