Componentize CaptivePortalBlockingPage [reland]
The componentization is trivial, as all of the work to clean the code
of //chrome-level deps has been done in preceding CLs.

This is a reland with the dependency of the security_interstitials
component on //components/wifi changed to be Mac/Win only, as
//components/wifi is implicitly limited to those two platforms. In an
esoteric way, adding the dependency unconditionally caused the
wifi_test target to be built on all platforms on the bots; that target
also is implicitly Mac/Win only and fails to link on other platforms.

Bug: 1025059
Change-Id: Ib4af301fe144b9530e1d32b5c55f8fda3c4cfe88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1953733
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723748}