[Cast MRP] Remove unwanted object clean-up.
The done_callback_, when executed, might have the side effect of
freeing the object executing it. This is because the callback refers
to the containing object and might be the last reference to it.

This CL removes unwanted state clean-up to avoid conflicting with the
object being released.

Bug: 1031653
Change-Id: Id335328ea297e5e3d772186f8e548f342025c1c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1961186
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Armando Miraglia <armax@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723763}