Add support for downloading PluginVM images from Drive.
Previously, PluginVmImageManager relied solely on the DownloadService
class to download the image. Now, if the file url is determined to
point to a Drive file, the PluginVmDriveImageDownloadService class
is used instead, which relies on drive::DriveService to make the API
calls and download the file.

Test: unit_tests
--gtest-filter="PluginVmUtilTest.DriveLinkDetection:PluginVmUtilTest.DriveLinkIdExtraction"

Bug: 1001463
Change-Id: Idf64eb5c70e6fac22c0efb947c6e5f945bb5faf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1936291
Commit-Queue: Muhamed Parić <muhamedp@google.com>
Reviewed-by: Olya Kalitova <okalitova@chromium.org>
Reviewed-by: Alex Oldemeier <aoldemeier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723768}