Removes the TODO to extend ambient auth tests for guest sessions in ChromeOS.
Ambient authentication is when the browser delegates the authentication
to the OS and the OS uses the default credentials to authenticate.
The guest sessions in ChromeOS by default were designed not to have
any state sharing with existing profiles and any default credentials.

Therefore, the TODO is removed and not valid.

Bug: 1030624
Change-Id: Ifbdde6954bd0f9bb4294cb241dae3995011ef6d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962188
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Rohit Agarwal <roagarwal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723829}