cbui views: admit AppMenuButton is a MenuButton
In particular, AppMenuButton uses MenuButtonController, which correctly
imposes the popup button role on it; the GetAccessibleNodeData override
in ToolbarButton unhelpfully changes the role back to a regular button.

As it happens, Button::GetAccessibleNodeData() already knows how to
configure both the role and the default action verb, so don't do that
in ToolbarButton.

The remaining code in ToolbarButton::GetAccessibleNodeData, which
sets HasPopup if there is a configured model, also seems suspect -
any button which has a menu should already be using MenuButtonController.
However, I am not confident enough that that actually obtains to
delete that case from ToolbarButton.

Bug: 954701
Change-Id: Ie2d5a364edef6298bc82681c7a4caff78c693f1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962665
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723891}