win: Collect indirectly-referenced memory in child process crash reports
Normally, a minidump contains snapshots of stack memory.
"Indirectly-referenced memory" is portions of the heap that appear to be
referenced by the stack. Collecting indirectly-referenced memory
inflates the size of a crash report, although not as much as a
full-memory dump would.

On Windows, it was intended to collect indirectly-referenced memory on
all channels except for stable. This was intended to function in all
process types. However, due to a bug, it was only effective in browser
processes.

This change enables the capture of indirectly-referenced memory in all
process types on Windows, except on the stable channel.

Bug: 1033039

Change-Id: Ib83e4a867142ded3eea01f5cdc7223de762358f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962645
Commit-Queue: Mark Mentovai <mark@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723895}