Android: Add Monochrome32 and Trichrome32 build targets
This change adds a pure 32-bit version of Monochrome and Trichrome to
the 64-bit build configuration. Building these targets generates a
product equivalent to building the targets on a 32-bit config. However,
by adding these, we have the option of building all Monochrome and
Trichrome products on a single 64-bit builder, and, we share the
compilation work (as opposed to the obsolete process of "APK merging"
parts from a 32-bit build into 64-bit products).

The targets added here are the public versions.

Bug: 1027573
Change-Id: Id5ffd698721077e8fffb1f8206adc0b7b1fce212
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960575
Reviewed-by: Eric Stevenson <estevenson@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Christopher Grant <cjgrant@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723897}