Define linux-wpt-fyi-rel builder as part of setting up an optional tryjob.
This is the second half of crrev.com/c/1960691, duplicating some
details from that cl below.
---
The goal here is to add a manually-triggered/optional tryjob to the CQ
that will run the WPT test suite.

A resource request was filed in crbug.com/1031562 and it was confirmed
there is enough capacity for this in the existing GCE pool.

There is already an existing waterfall builder at
https://ci.chromium.org/p/chromium/builders/ci/linux-wpt-fyi-rel and
the new tryjob is intended to mirror this builder and tester setup.
---

Bug: 937369
Change-Id: I2ea9b0830799a8f4de687f40fa6ef4839f8d5923
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962647
Commit-Queue: Luke Z <lpz@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723953}