Make kZeroWidthNonJoinerCharacter as a grapheme break.
When using Myanmar keyboard on Windows, some characters are preceded
by a kZeroWidthNonJoinerCharacter so it doesn't collapse and join with
the previous character to form a new grapheme cluster while there is a
composition in progress. When IME tried to replace the character under
active composition, insertText command in blink used visible position
instead of logical character sequence which snapped the positions at the
end of the composition, resulting in a duplicate insertion. During visible
position creation, the |IsGraphemeBreak| check returns false when the codepoint
is kZeroWidthNonJoinerCharacter so it skips this position.
Added kZeroWidthNonJoinerCharacter as a grapheme break so the positions can be
properly used to replace the composed character by the IME.

Bug: 1027695

Change-Id: I427a63afaa8b37babd76554de7b266050c44ebf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960986
Commit-Queue: Anupam Snigdha <snianu@microsoft.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723960}