[HaTS] Don't show HaTS when incognito mode disabled
When incognito mode is disabled, creating an off-the-record profile will
run into a check to make sure incognito mode is enabled. If not, it
leads to immediate crash. Off-the-record profile is needed for showing
the HaTS web dialog. So, this CL adds a check to not show HaTS bubble
when incognito mode is disabled.

BUG=1032459

Change-Id: I834f776c1239c8f163d6e7607767ba59f5a1d1c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1961212
Commit-Queue: Robert Liao <robliao@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723972}