Use TestSyncService in OsSyncHandlerTest instead of GMock
I find GMock a little hard to read, so switch the test to using the
existing TestSyncService class instead.

Change the tests for navigating away / closing settings to start
with the feature disabled and simulate the user clicking "Turn on".
This is more like the flow we care about (not enabling OS sync
before the user is done setting it up).

Bug: 1013466
Test: unit_tests
Change-Id: I58b5a6cbd58ed9ebbf14fff9e40353e3f4b17185
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962770
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723977}