Update Disabled state for TestControllerInputRegistered
Rather than disabling TestControllerInputRegistered for all runtimes,
this "disables" (by way of an early return) just the OpenVr runtime,
which is the one that is flaky.

To do this, it also adds a macro "WEBXR_VR_DISABLE_TEST_ON" which can
be used to do something similar in the future when only one runtime
should be disabled.

Bug: 1033087
Change-Id: Ieda1b084a8c888333bf1ac2875e911e3d537e832
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1963096
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723987}