Fix cursor appearance on environments that do not set cursor settings
On environments that do not have GTK settings for the cursor theme/size such as
KDE, we would set the cursor theme name to "", and the cursor size to 0. The
theme would be fine, but the cursor could appear too small on HiDpi
environments. The solution is to avoid making any XcursorSet* calls if the
settings are not present.

BUG=1021819
R=sky

Change-Id: I993e356358981abe22649faa673f234c526ff654
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962710
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724060}