media/gpu: return Optional<Fourcc> from Fourcc::FromV4L2PixFmt()
Invalid Fourccs are dangerous by nature: they can be passed around just
like valid values and create bugs that manifest in a completely
different part of the code. To solve this, we plan to make it impossible
to create an invalid Fourcc instance.

This CL advances towards this goal by making Fourcc::FromV4L2PixFmt()
return Optional<Fourcc>. If a value exists, it is guaranteed to be
valid, and if the conversion failed then nullopt is returned. This
forces the caller to check the returned value and will crash immediately
if the Fourcc is access in an attempt to pass the value around. All
callers of this method are also updated accordingly, with the required
checks added (these checks should have been there anyway).

BUG=1031082
BUG=b:143498574
TEST=VDAtests passing on Hana.

Change-Id: Ice2aa403b4e2e53ea6a0be2d0d165b7a2337ad7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1957195
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724100}