AW UMA: add DriveMetricsProvider
This adds DriveMetricsProvider to WebView's list of MetricsProviders. At
the moment, this is hardcoded to log `has_seek_penalty == false`, but
this registers the provider anyway for the sake of consistency with
Chrome, and in case the implementation changes in the future to log
other values.

Bug: 1008534
Test: run_webview_instrumentation_test_apk -f AwMetricsIntegrationTest#*
Change-Id: I8e9b0c91bab1dffa5b6c13ccc323ae7bfe46eef9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960476
Reviewed-by: Tao Bai <michaelbai@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724142}