Integration test for propagation of error dismiss v2
Test for CL:1943777 behavior.

The background page for FilesApp stores state for file operations in
progress, as well as error states. This information is propagated to
all the Files App foreground windows during their lifetime, and also
propagated to any newly created foreground windows.

The background progress center has 2 callback functions for changing
state - one for cancel of an operation in progress, the other for
dismissal of an error state.

This test triggers an error state, simulates a click on the 'dismiss'
button on the error panel that is displayed in the foreground window. It
then closes the foreground window and opens a new one to check the error
panel is not displayed. Detection of the lack of an error panel is done
by triggering a copy operation that never ends and checking for that
panel. If the error had persisted, it would be present when the copy
panel appeared.

Bug: 990919
Tests: browser_tests --gtest_filter="*transferDismissedErrorIsRemembered"
Change-Id: I644f8b8d5ae10080f8fcf6dc7813bc8afdc92e83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1961835
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724156}