[iOS] Refactor LogoVendor provider to pass webState (1/2)
To remove the use of the deprecated new_tab_page_uma API (see bug) in
provided objects, the LogoVendor provider API needs to be able to pass
a webState along with the BrowserState into the provider.

This CL adds this API; the child CL removes the API that lacks the
webstate parameter. The downstream CL to use this new API will need to
land in between these two CLs.

Bug: 1032550
Change-Id: I430335912216744fa47c98cb9006884422a3e146
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962250
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724158}