Apply style adjustment to pseudo-elements without CSS rules
Before this patch, if there was no CSS rule for a pseudo-element, its
style wasn't adjusted. For example, if the parent was a grid or flex
container, the pseudo-element wasn't blockified.

For ::before and ::after this only mattered in getComputedStyle, since
they don't generate boxes by default. But this change can be more
important for ::marker, since it does generates a box even when no rule
is specified for it.

BUG=1033058

TEST=external/wpt/css/cssom/getComputedStyle-pseudo.html

Change-Id: I6452d4c81da5fb7f30d50f0df295be43299bc278
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962284
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#724172}