Desktop PWAs: Add window title after frame toolbar
Since minimal-ui was implemented, the web app frame toolbar occupies
the full width of the window caption bar, not just a small amount of
space on the right.

The window title needs to be layered above the titlebar so it can
receive hover messages and show a tooltip with the full application
title when the window is narrow compared to the application title
length.

Bug: 1026158
Change-Id: I89b592754c51af69305ddad9b05db0a3400250e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1958301
Commit-Queue: Eric Willigers <ericwilligers@chromium.org>
Reviewed-by: Bret Sepulveda <bsep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724181}