Fix spelling errors in pixel_expectations
Follow up to: https://chromium-review.googlesource.com/c/chromium/src/+/1961604/
This fixes typos of the comments for the suppressions updated there.

As a note, in the original patch some of the expectations were also updated
to have new bug numbers. This is because either older bugs were generic
and since closed. Or the form of error they demonstrate is different now.

TBR=ynovikov@chromium.org
TEST=pixel_tests

Bug: 1008450, 1033113, 1033121
Change-Id: Ia808a2bc9e28fd167fb000b523ee81155ad7268f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1964234
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724194}