cbui crostini: remove CrostiniInstallerView DialogDelegate overrides
This change removes overrides:
* CrostiniInstallerView::GetDialogButtons
* CrostiniInstallerView::GetDialogButtonLabel

In favor of the new explicit setters, and adds a new function
(UpdateDialogButtonsAfterStateChange) to indicate explicitly to DialogDelegate
that the buttons have changed. This is desirable here because this delegate has
many different states that it goes through during its lifecycle and making it
explicit when a change is intended to adjust the buttons clarifies what's going
on.

This change also renames the old methods to GetCurrentDialogButtons and
GetCurrentDialogButtonLabel, both to de-conflict with the existing methods on
DialogDelegate and to clarify that their return values can change over the
lifetime of the dialog, which is not usually true for overrides of these
methods.

Bug: 1011446
Change-Id: Ia420b419aea2f6c8b0576af29f52018b3d5b25aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960867
Reviewed-by: Timothy Loh <timloh@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724199}