[Web Payment][Android] Payment app factory delegate.
Before this patch, PaymentRequestImpl.java was responsible for receiving
a list of PaymentApp.java instances, which it then queried for a list of
PaymentInstrument.java instances. This created complexity in
PaymentRequestImpl.java file, which grew very long.

This patch moves the logic for interaction with PaymentApp.java and
PaymentInstrument.java instances into PaymentAppFactory.java. The
PaymentRequestImpl.java class now implements the
PaymentAppFactoryDelegate.java interface.

The PaymentAppFactory.java class takes an instance of
PaymentAppFactoryDelegate.java interface that provides data necessary
for creation of payment apps and also receives the list of available
payment apps.

The PaymentAppFactoryDelegate.java interface is using the new method and
parameter naming convention of calling everything that can handle
payment an "app", including Autofill credit cards. Everything that can
create an "app" is a "factory". The interfaces and implementations of
PaymentApp.java and PaymentInstrument.java have not been renamed yet to
avoid breaking the Google Chrome build.

The implementers of the interfaces use the parameter naming conventions
of the interfaces. Unfortunately this results in a bit of confusion for
now, because PaymentAppFactory.java uses PaymentAppFactoryDelegate.java
with new naming conventions and implements PaymentAppCreatedCallback and
InstrumentsCallback interfaces that still use the old naming
conventions.

After this patch, PaymentRequestImpl.java implements
PaymentAppFactoryDelegate.java and receives a list of
PaymentInstrument.java instances from PaymentAppService.java, which
reduces the complexity in PaymentRequestImpl.java class. The class is
now 50 lines shorter and contains more @Overrides for accessors in
PaymentAppFactoryDelegate.java instead of business logic, which has been
moved to PaymentAppFactory.java.

Design: https://bit.ly/cross-platform-pay-app-factory

Bug: 1022512
Change-Id: I828071cc18b6b5c5f3fa03d1864b91b9adab872d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1950748
Reviewed-by: Danyao Wang <danyao@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724207}