[Autofill Assistant] Send previous selection of objects
Previously when setting items the initial object was automatically
selected in the view in Java. Now this initial selection has been moved
to native (still in the view).
This means, that resetting a selection now also falls into the realm of
native.

Adding tests to verify the behaviour.

Bug: b/144005336
Bug: b/144681042
Change-Id: I5b684d16df0aeab36fdc4d738662c5292e8a99e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954393
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#724208}