MacPWAs: Fix aberrant user manager (and crash)
The switches::kProfileDirectory argument is supposed to be a relative
path. The app shim specifies AppShimController::Params::profile_dir,
which used to be a relative path but now is absolute (maybe it
shouldn't be?).

When an invalid relative path is specified, the the logic in
StartupBrowserCreator::ProcessCmdLineImpl interprets this the same way
as there having been no last active profile, and opens the user
manager.

Just stop specifying switches::kProfileDirectory, since the idea is
going away.

Also retain the result from OpenApplicationWithPath, as it was causing
use-after-free errors observed in debugging this.

Bug: 1032242
Change-Id: Iaf8ec1131a5893387bd54a1a686c55828f63ba6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962613
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724224}