Move calls of |g_browser_process| to the main/UI thread
A previous commit makes calls to |g_browser_process| from threads
other than the main thread  [1]. The BrowserProcess class is not
thread safe, so pass the locale from the main thread to the threads
that need it.

Add a DCHECK_CURRENTLY_ON() to FetchCapabilities() in
arc_print_service.cc to ensure that the call of |g_browser_process|
is on the main thread.

[1] crrev.com/722201

Bug: 890904
Change-Id: I6b021fba9ba03d157a9ca33bd7c812a36cac7f74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1963090
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Daniel Hosseinian <dhoss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724244}