[ios] Pack localisation for MaterialTabs.bundle
During review of https://crrev.com/c/1960457, it was found that
only "en" localisation of MaterialComponents components where
packed. It turns out that one component "MaterialTabs.bundle"
has localisation in other locales (this was not visible in the
app as the localisation is used for accessibility labels only).

This CL packs those localisation that are shared with the rest
of supported languages by Chrome on iOS.

Bug: 1017165
Change-Id: I0c25f24357c6867dcb08f518a161a84f3ccd7fdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1964611
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724300}