Refactor PageInfoController#show: replace Tab with WebContents
Old behavior:
The PageInfo dialog can only be shown for a tab-based WebContents.

New behavior:
The PageInfo dialog can be shown for non-tab-based WebContents as well.
This would be useful if we want to show the dialog for
PaymentHandler UI, which is based on a WebContents instead of a Tab.

Change:
Refactor PageInfoController#show to remove the tab dependency

Side change:
Remove context as a member of PageInfoController. Instead, get context
from mWindowAndroid when needed. This is because WebContents'
reparenting could cause context and mWindowAndroid to be mismatched.

Bug: 999196
Change-Id: I9d43a9892c812105191922edd568a84679c3cfac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1944649
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724301}