Break autofill assistant's dep on compositor
This patch removes autofill's dependency on the compositor's
EventFilter class. In practice, the only reason this class was used
was to have access to the onTouchEvent method which can be achieved
via Android's GestureDetector. I opted to leave the name of the class
the same as it is still filtering events and never calls super for
the overridden methods.

Change-Id: If03faceef99e365915b41bf440a9fdc2d164fe7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962988
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724319}