Support compressed .json grit files, compress chromevox msgs (Take 4)
This provides support for type="chrome_messages_json_gzip" in
grit/build.py, which compresses the output files (which should have
.gz appended to the name, e.g. messages.json.gz).

This makes it fairly simple to convert an extension (e.g. chromevox)
to use compressed l10n files.

Original CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1913610
This is identical to "Take 3" which was reverted because of missing
support for compressed locale files in comoponent extensions, fixed in:
https://chromium-review.googlesource.com/c/chromium/src/+/1956224

For changes reviewed in the original CL:
TBR=agrieve@chromium.org
TBR=dtseng@chromium.org
TBR=rdevlin.cronin@chromium.org

Bug: 1023568
Change-Id: I177fe5470ad8e88d279e2d8de3938581331b51e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1963190
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724324}